refactor: remove parallaxActSteps and related sections for a cleaner architecture
